Fix split
