+++ config/splat.us.stno1.yaml
@@ -46,7 +46,7 @@ segments:
       - [0xCFC, .data, st_debug]
       - [0xD7C, data]
       - [0x1508, data, unk_3B5E0]
-      - [0x1550, data]
+      - [0x1540, data, unk_3CB34]
       - [0x1570, .data, unk_3D8AC]
       - [0x15F4, data]
       - [0x1A2C, .data, d_prize_drops]
@@ -102,7 +102,7 @@ segments:
       - [0x38008, c, e_chair]
       - [0x381E8, c, unk_381E8]
       - [0x3B5E0, c, unk_3B5E0]
-      - [0x3CB34, c, unk_3CB34]
+      - [0x3C598, c, unk_3CB34]
       - [0x3D8AC, c, unk_3D8AC]
       - [0x3E880, c, e_wall_chicken_secret]
       - [0x3FA34, c, unk_3FA34]
