+++ src/st/no1/unk_3B5E0.c
@@ -9,7 +9,6 @@ INCLUDE_ASM("st/no1/nonmatchings/unk_3B5E0", func_us_801BB7B8);
 
 extern s16 D_us_80181508[][3];
 extern s16 D_us_8018151C[][3];
-extern u8 D_us_80181540[];
 
 extern s32 D_psp_08C62AA4;
 
@@ -413,154 +412,3 @@ void func_us_801BBD90(Entity* self) {
 }
 
 STATIC_PAD_RODATA(4);
-
-void func_us_801BC598(Primitive* prim) {
-    Primitive* otherPrim;
-    u8 tempVar;
-
-    switch (prim->next->u2) {
-    case 0:
-        prim->tpage = 0x1A;
-        prim->clut = 0x1B0;
-        prim->u0 = 0x70;
-        prim->u1 = 0x7F;
-        prim->u2 = prim->u1;
-        prim->u3 = prim->u2;
-        prim->v0 = 0x20;
-        prim->v1 = prim->v0;
-        prim->v2 = 0x2F;
-        prim->v3 = prim->v2;
-        prim->next->b3 = 0x80;
-        if (prim->next->r3 & 0x10) {
-            if (prim->next->r3 & 1) {
-                LOH(prim->next->tpage) = 0xA00;
-            } else {
-                LOH(prim->next->tpage) = 0x600;
-            }
-        } else {
-            if (prim->next->r3 & 1) {
-                LOH(prim->next->tpage) = -0x100;
-            } else {
-                LOH(prim->next->tpage) = 0x100;
-            }
-        }
-        LOH(prim->next->r2) = 0xE;
-        LOH(prim->next->b2) = 0x10;
-        prim->priority = 0xF0;
-        prim->drawMode = DRAW_TPAGE2 | DRAW_TPAGE | DRAW_UNK02 | DRAW_TRANSP;
-        prim->next->x3 = 0x10;
-        prim->next->u2 = 1;
-        /* fallthrough */
-    case 1:
-        if (prim->next->r3 & 1) {
-            prim->next->x1 = g_CurrentEntity->posX.i.hi - 0xC;
-        } else {
-            prim->next->x1 = g_CurrentEntity->posX.i.hi + 0xC;
-        }
-        if (prim->next->r3 & 0x10) {
-            prim->next->y0 = g_CurrentEntity->posY.i.hi - 0x25;
-        } else {
-            prim->next->y0 = g_CurrentEntity->posY.i.hi + 0x23;
-        }
-        UpdateAnimation(D_us_80181540, prim);
-        UnkPrimHelper(prim);
-        if (Random() % 8 == 0) {
-            otherPrim = g_CurrentEntity->ext.et_801BBD90.unk7C;
-            otherPrim = FindFirstUnkPrim2(otherPrim, 2);
-            if (otherPrim != NULL) {
-                UnkPolyFunc2(otherPrim);
-                otherPrim->x0 = prim->next->x1;
-                otherPrim->y0 = prim->next->y0;
-                otherPrim->g1 = (Random() & 1) + 1;
-                otherPrim->g2 = (Random() & 3) + 1;
-                otherPrim->next->u2 = 4;
-                otherPrim->next->r3 = prim->next->r3;
-            }
-        }
-        prim->next->x3--;
-        if (!prim->next->x3) {
-            prim->next->u2 = 0;
-            prim->p1 = 0;
-            prim->p2 = 0;
-            UnkPolyFunc0(prim);
-        }
-        break;
-
-    case 4:
-        prim->type = PRIM_TILE;
-        prim->u0 = 1;
-        prim->v0 = 1;
-        prim->b0 = 0xC0;
-        prim->r0 = 0x40;
-        prim->g0 = 0x40;
-        prim->priority = 0xC0;
-        prim->drawMode = DRAW_TPAGE2 | DRAW_TPAGE | DRAW_TRANSP;
-        prim->next->x3 = 0x20;
-        prim->next->u2 = 5;
-        /* fallthrough */
-    case 5:
-        if (prim->next->r3 & 0x10) {
-            prim->y0 -= prim->g1;
-        } else {
-            prim->y0 += prim->g1;
-        }
-        tempVar = prim->g2;
-        if (g_Timer % tempVar == 0) {
-            if (prim->next->r3 & 1) {
-                prim->x0++;
-            } else {
-                prim->x0--;
-            }
-        }
-        prim->next->x3--;
-        if (!prim->next->x3) {
-            prim->type = PRIM_GT4;
-            prim->next->u2 = 0;
-            UnkPolyFunc0(prim);
-        }
-        break;
-    }
-}
-
-void func_us_801BC9A8(Primitive* prim) {
-    s16 posY;
-
-    switch (prim->next->u2) {
-    case 0:
-        prim->tpage = 0xE;
-        prim->clut = 0x4A;
-        prim->u0 = 0xD0;
-        prim->u1 = prim->u0 + 0x1F;
-        prim->u2 = prim->u0;
-        prim->u3 = prim->u1;
-        prim->v0 = 0xC0;
-        prim->v1 = prim->v0;
-        prim->v2 = prim->v0 + 0xF;
-        prim->v3 = prim->v2;
-        prim->priority = 0x70;
-        prim->drawMode = DRAW_TPAGE | DRAW_COLORS | DRAW_UNK02 | DRAW_TRANSP;
-        LOH(prim->next->r2) = 0x20;
-        LOH(prim->next->b2) = 0x10;
-        prim->next->b3 = 0x30;
-        prim->next->u2 = 1;
-        /* fallthrough */
-    case 1:
-        prim->next->x1 = g_CurrentEntity->posX.i.hi;
-        prim->next->y0 = g_CurrentEntity->posY.i.hi + 0x24;
-        UnkPrimHelper(prim);
-        prim->drawMode = DRAW_TPAGE | DRAW_COLORS | DRAW_UNK02 | DRAW_TRANSP;
-        if (!g_CurrentEntity->velocityY || g_Timer % 2) {
-            prim->drawMode =
-                DRAW_TPAGE | DRAW_COLORS | DRAW_UNK02 | DRAW_TRANSP;
-        } else {
-            prim->drawMode = DRAW_HIDE;
-        }
-        posY = g_Tilemap.scrollY.i.hi + prim->y2;
-        if (posY > 0x810) {
-            prim->drawMode = DRAW_HIDE;
-        }
-        if (posY > 0x610 && posY < 0x650) {
-            prim->drawMode = DRAW_HIDE;
-        }
-    }
-}
